Change security-proxy so that sec-username and sec-roles always are removed.
Added new header provider in security-proxy that checks the currently logged in user and if that user is one of the "special"/"trusted" users then it will map

imp-username -> sec-username
imp-roles -> sec-roles

Allowing trusted users to impersonate another user.  This is required for the extractorapp to enforce the security on the end user.