security-proxy - ignores non HTTPS request check
The security-proxy can bypass CAS authentication in case of specific
user-agents and a HTTP basic-auth is provided. But the filter discards
requests that are not targetted onto https.

This adds an option to ignore if the request targets or not an HTTPS
server (anyway, the geOrchestra instances spread over multiple servers
should be bound over a trusted network, so the risk of sniffing
passwords *should* be mitigated).