security-proxy - improving custom authentication filter
Adds the ability to try a bind onto the OpenLDAP in case of failure on
the Active Directory. This should address the bugs related to
server-to-server communications, and removing the whitelisting.

This code is supposed to be rennes-project specific.

TODO: there will still be some issues on sig-test instance

* SSL certificates mismatch while extractorapp -> geoserver
* there is still a whitelist when geofence -> geoserver, which
causes false positive result reguarding to the current code
modification.