security-proxy - Intercept "connection refused" errors
If underlying webapps are unavailable, an empty webpage is produced
(Content-length: 0), no log traces are produced (a printStackTrace() is
done anyway), and a 200 OK status code is returned.

This commit provides:
- A 503 "service unavailable" status code is returned instead
- A default status page for 503 (as it is currently done for 403 and 404
  statuses) is produced,
- A log trace is correctly (in the correct log file, not on stdout) is
  printed.

Note that the 503 error code can also be intercepted by the frontend
server, so that a custom error page is returned.