SP - fixing parameters passed in GET
Following the SP revamp, every parameters passed as in a form-encoded
POST were also passed as GET parameters to the security-proxified
webapps.

Fixing the logic so that parameters present in the URI are copied.

Also, there was a potential issue with UTF-8 accented characters, that
this commit shall address.

Tests:
* Unit tests added
* Runtime tested on Jetty (using a simple controller)
* Runtime tested on Tomcat (sandbox.georchestra.org)