SP - configure the http request timeout
When a user uploads a big file on a rather bad connection, this can take
some time. Currently, the request is made onto the SP in an asynchronous
fashion (to avoid getting the whole response in memory before giving it
back to the client). There is a timeout set to 5 minutes in the code
that limits the time for the client being able to send data to the
proxified server, regardless of other timeouts set on frontend proxies.

If the frontend sets a timeout to 20 minutes, then we should also be
able to allow the query to take the same amount of time when getting
through the SP.

Note: this revealed another issue: In case of reaching the "query
timeout",we need to wait for the end of the data from the client being
sent before giving a bad status.  Unfortunately, it does not seem
possible with Servlets to interrupt / close the request, even if we know
that the query failed.

Tests:

* No tests added, this case being pretty hard to set up a testing
  scenario
* mvn clean verify -Pit ok though
* runtime tested on a dev env with a low timeout set & bandwidth
  trickled