sp - get a hand on the raw urlencoded post body
Roughly inspired from https://howtodoinjava.com/servlets/httpservletrequestwrapper-example-read-request-body/

See https://github.com/georchestra/georchestra/pull/3070#issuecomment-639029518 for the motivation

This makes use of a servlet filter to wrap the request in a custom
wrapper, in case it is:

1. an HttpServletRequest
2. a POST request
3. a URL-encoded POST request

This allows us to stay the closest possible from the origin POST request
sent by the browser, saving it onto the request as a custom attribute.

This also allows to factor some bits on how the POST requests are
managed in the main Proxy.java class.

tests: mainly runtime-tested on a local compo with GeoServer.