sp - explicitely defines supported methods
This is to be able to have the OPTIONS requests managed by the SP (and
the underlying proxified webapps).

Normally we would like to leave the method parameter empty so that it
defaults to the full list of supported operations. But since Spring
introduces a special treatment for OPTIONS requests:
https://github.com/spring-projects/spring-framework/blob/v4.3.22.RELEASE/spring-webmvc/src/main/java/org/springframework/web/servlet/mvc/annotation/ServletAnnotationMappingUtils.java#L48

We are willing to skip the test at the line mentioned above, and
be able to reach our applicative code afterwards. the result of
the OPTION request will be left to the proxified webapp.

Tests: runtime tested