sp - harmonize /whoami endpoint output with the gateway (#4210)
Autowire the usersApi bean to the Proxy class so that we can resolve a
GeorchestraUser object from the connected user. This allows to get the
same java object as the one being returned by the same endpoint from the
Gateway.

Tests:
* utests adapted
* runtime tested with testadmin:testadmin + anonymous user (blank call
  to /whoami)

Note: I'd have liked to leave the JSON serialization to spring, as it is
done into the Gateway, but could not make it. This would have avoided
the n calls to JSONObject.put, but sounds acceptable to me for the time
both SP and Gateway will coexist.